requirements.txt
google-genai>=0.8.0
markdown>=3.5
pytz>=2024.1
